renamed files according to code convention and should be tested
