+++ Server/RequestProcessing/new_user_request.cpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "NewUserRequest.h"
+#include "new_user_request.h"
