+++ Server/Server.vcxproj.filters
@@ -4,15 +4,6 @@
     <ClCompile Include="RequestProcessing\add_friend_request.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="RequestProcessing\AbstractRequest.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="AsyncServer.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="DAL\client.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="DAL\dal.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
@@ -22,19 +13,25 @@
     <ClCompile Include="Logger\logger.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="RequestProcessing\LoginRequest.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
     <ClCompile Include="main.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-    <ClCompile Include="RequestProcessing\NewUserRequest.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="..\Parser&amp;Structs\parser.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="ServerThread.cpp">
+    <ClCompile Include="server_thread.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="async_server.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="RequestProcessing\new_user_request.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="RequestProcessing\login_request.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="RequestProcessing\abstract_request.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -48,16 +45,10 @@
     <ClInclude Include="Logger\logger.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="DAL\Client.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="..\Parser&amp;Structs\parser.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="Exceptions\user_not_found.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="RequestProcessing\NewUserRequest.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="..\Parser&amp;Structs\request_types.h">
@@ -66,18 +57,21 @@
     <ClInclude Include="RequestProcessing\friendship_request.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="RequestProcessing\LoginRequest.h">
+    <ClInclude Include="RequestProcessing\new_user_request.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="RequestProcessing\login_request.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
-    <QtMoc Include="RequestProcessing\AbstractRequest.h">
+    <QtMoc Include="server_thread.h">
       <Filter>Header Files</Filter>
     </QtMoc>
-    <QtMoc Include="AsyncServer.h">
+    <QtMoc Include="async_server.h">
       <Filter>Header Files</Filter>
     </QtMoc>
-    <QtMoc Include="ServerThread.h">
+    <QtMoc Include="RequestProcessing\abstract_request.h">
       <Filter>Header Files</Filter>
     </QtMoc>
   </ItemGroup>
